Jmol.___JmolVersion="15.2.53" // P also 14.32.53
(into Jmol 14 only from precision)
bug fix: PWmatReader fails to adjust atom properties for skipped atoms
bug fix: PWmatReader fails to retain properties after MODELKIT ADD ...
bug fix: adding an atom removes atom properties
bug fix: filter "*.CA" does not work

+++ src/org/jmol/adapter/readers/cif/MMCifReader.java
@@ -26,20 +26,20 @@ package org.jmol.adapter.readers.cif;
 import java.util.Hashtable;
 import java.util.Map;
 
+import org.jmol.adapter.smarter.Atom;
+import org.jmol.adapter.smarter.Structure;
+import org.jmol.api.JmolAdapter;
+import org.jmol.c.STR;
+import org.jmol.util.BSUtil;
+import org.jmol.util.Logger;
+import org.jmol.util.SimpleUnitCell;
+
+import javajs.util.BS;
 import javajs.util.Lst;
 import javajs.util.M4d;
 import javajs.util.P3d;
 import javajs.util.PT;
 import javajs.util.SB;
-
-import org.jmol.adapter.smarter.Atom;
-import org.jmol.adapter.smarter.Structure;
-import org.jmol.api.JmolAdapter;
-import org.jmol.c.STR;
-import javajs.util.BS;
-import org.jmol.util.BSUtil;
-import org.jmol.util.Logger;
-import org.jmol.util.SimpleUnitCell;
 
 
 /**
@@ -604,7 +604,7 @@ public class MMCifReader extends CifReader {
           xyz = field;
           break;
         default:
-          m[p] = (double) parseDoubleStr(field);
+          m[p] = parseDoubleStr(field);
           ++count;
         }
       }
